WCM-545 prevent null pointer exception in tests

+++ content-targeting-service/src/main/java/com/liferay/content/targeting/service/impl/UserSegmentLocalServiceImpl.java
@@ -489,6 +489,10 @@ public class UserSegmentLocalServiceImpl
 		AssetCategory assetCategory =
 			assetCategoryLocalService.fetchAssetCategory(assetCategoryId);
 
+		if (assetCategory == null) {
+			return;
+		}
+
 		try {
 			assetCategoryLocalService.deleteCategory(assetCategory);
 		}
